Update README.md
Powerco is a major gas and electricity utility that supplies corporate, SME (Small & Medium enterprises), and residential customers. The power liberalization of the energy market in Europe has led to significant customer churn, especially in the SME segment. They have partnered with BCG to help diagnose the source of churning SME customers.